Conflicts:
	blackboardvc-portlet-webapp/src/main/java/org/jasig/portlet/blackboardvcportlet/mvc/sessionmngr/SessionCreateEditController.java

+++ blackboardvc-portlet-webapp/src/main/java/org/jasig/portlet/blackboardvcportlet/service/impl/SessionServiceImpl.java
@@ -2,7 +2,9 @@ package org.jasig.portlet.blackboardvcportlet.service.impl;
 
 import java.io.File;
 import java.io.IOException;
+import java.util.Iterator;
 import java.util.LinkedHashSet;
+import java.util.List;
 import java.util.Set;
 
 import javax.activation.DataHandler;
@@ -13,14 +15,18 @@ import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.FilenameUtils;
 import org.jasig.portlet.blackboardvcportlet.dao.ConferenceUserDao;
 import org.jasig.portlet.blackboardvcportlet.dao.MultimediaDao;
+import org.jasig.portlet.blackboardvcportlet.dao.PresentationDao;
 import org.jasig.portlet.blackboardvcportlet.dao.SessionDao;
+import org.jasig.portlet.blackboardvcportlet.dao.UserSessionUrlDao;
 import org.jasig.portlet.blackboardvcportlet.dao.ws.MultimediaWSDao;
 import org.jasig.portlet.blackboardvcportlet.dao.ws.PresentationWSDao;
 import org.jasig.portlet.blackboardvcportlet.dao.ws.RecordingWSDao;
 import org.jasig.portlet.blackboardvcportlet.dao.ws.SessionWSDao;
 import org.jasig.portlet.blackboardvcportlet.data.ConferenceUser;
 import org.jasig.portlet.blackboardvcportlet.data.Multimedia;
+import org.jasig.portlet.blackboardvcportlet.data.Presentation;
 import org.jasig.portlet.blackboardvcportlet.data.Session;
+import org.jasig.portlet.blackboardvcportlet.data.UserSessionUrl;
 import org.jasig.portlet.blackboardvcportlet.security.ConferenceUserService;
 import org.jasig.portlet.blackboardvcportlet.service.SessionForm;
 import org.jasig.portlet.blackboardvcportlet.service.SessionService;
@@ -28,6 +34,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.oxm.XmlMappingException;
+import org.springframework.remoting.soap.SoapFaultException;
 import org.springframework.security.access.prepost.PostFilter;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Service;
@@ -38,7 +45,11 @@ import org.springframework.web.util.WebUtils;
 import org.springframework.ws.client.WebServiceClientException;
 
 import com.elluminate.sas.BlackboardMultimediaResponse;
+import com.elluminate.sas.BlackboardPresentationResponse;
 import com.elluminate.sas.BlackboardSessionResponse;
+import com.google.common.base.Function;
+import com.google.common.collect.ImmutableSet;
+import com.google.common.collect.Iterators;
 
 @Service
 public class SessionServiceImpl implements SessionService, ServletContextAware {
@@ -48,13 +59,27 @@ public class SessionServiceImpl implements SessionService, ServletContextAware {
 	private ConferenceUserDao conferenceUserDao;
     private SessionDao sessionDao;
     private MultimediaDao multimediaDao;
+    private PresentationDao presentationDao;
 	private SessionWSDao sessionWSDao;
 	private MultimediaWSDao multimediaWSDao;
 	private PresentationWSDao presentationWSDao;
 	private RecordingWSDao recordingWSDao;
 	private File tempDir;
+
+	private UserSessionUrlDao userSessionUrlDao;
 	
 	@Autowired
+	public void setUserSessionUrlDao (UserSessionUrlDao dao) {
+		this.userSessionUrlDao = dao;
+	}
+	
+	
+	@Autowired
+	public void setPresentationDao(PresentationDao presentationDao) {
+        this.presentationDao = presentationDao;
+    }
+
+    @Autowired
 	public void setMultimediaDao(MultimediaDao multimediaDao) {
         this.multimediaDao = multimediaDao;
     }
@@ -110,7 +135,29 @@ public class SessionServiceImpl implements SessionService, ServletContextAware {
     public Set<ConferenceUser> getSessionChairs(Session session) {
         return new LinkedHashSet<ConferenceUser>(sessionDao.getSessionChairs(session));
     }
+    
+    @Override
+	@PreAuthorize("hasRole('ROLE_ADMIN') || hasPermission(#session, 'view')")
+    public boolean isSessionParticipant(Session session, ConferenceUser user) {
+    	ConferenceUser userFromDB = conferenceUserDao.getUser(user.getUserId());
+    	return (sessionDao.getSessionChairs(session).contains(userFromDB)
+    			|| sessionDao.getSessionNonChairs(session).contains(userFromDB));
+    }
 
+	@PreAuthorize("hasRole('ROLE_ADMIN') || hasPermission(#session, 'view')")
+	@Transactional
+	public String getOrCreateSessionUrl(ConferenceUser user, Session session) {
+		//check for the url in the db
+		UserSessionUrl url = userSessionUrlDao.getUserSessionUrlsBySessionAndUser(session, user);
+		
+		if(url == null) {
+			//if null then create a user's session url via web service call
+			String urlString = sessionWSDao.buildSessionUrl(session.getBbSessionId(), user);
+			//save to the database
+			url = userSessionUrlDao.createUserSessionUrl(session, user, urlString);
+		}
+		return url.getUrl();
+	}
     /**
      * A user needs "edit" to view the set of session non chairs but we don't want the call to fail
      * if they only have "view" permission. So we pre-auth them with view and then filter all
@@ -140,6 +187,7 @@ public class SessionServiceImpl implements SessionService, ServletContextAware {
     public Session getSession(long sessionId) {
         return this.sessionDao.getSession(sessionId);
     }
+    
 
     /*
      * Not rolling back for WS related exceptions so the work done "so far" is still persisted to the database in
@@ -151,21 +199,41 @@ public class SessionServiceImpl implements SessionService, ServletContextAware {
     public void removeSession(long sessionId) {
         final Session session = this.sessionDao.getSession(sessionId);
         
+        final Set<Long> bbMultimediaIds = getBlackboardMultimediaIds(session);
         final Set<Multimedia> multimedias = this.sessionDao.getSessionMultimedias(session);
         for (final Multimedia multimedia : multimedias) {
-            //Un-link multimedia file from session
-            this.multimediaWSDao.removeSessionMultimedia(session.getBbSessionId(), multimedia.getBbMultimediaId());
-            this.sessionDao.deleteMultimediaFromSession(sessionId, multimedia);
-            
-            //Delete multimedia file from repository
-            this.multimediaWSDao.removeRepositoryMultimedia(multimedia.getBbMultimediaId());
-            this.multimediaDao.deleteMultimedia(multimedia);
+            removeMultimediaFromSession(session, bbMultimediaIds, multimedia);
         }
         
-        //TODO delete presentation
+        this.deletePresentation(session.getSessionId());
         
         this.sessionWSDao.deleteSession(session.getBbSessionId());
         this.sessionDao.deleteSession(session);
+    }
+
+    private void removeMultimediaFromSession(Session session, Set<Long> bbMultimediaIds, Multimedia multimedia) {
+        //Un-link multimedia file from session
+        if (bbMultimediaIds.contains(multimedia.getBbMultimediaId())) {
+            this.multimediaWSDao.removeSessionMultimedia(session.getBbSessionId(), multimedia.getBbMultimediaId());
+        }
+        this.sessionDao.deleteMultimediaFromSession(session, multimedia);
+
+        //Delete multimedia file from repository
+        try {
+            this.multimediaWSDao.removeRepositoryMultimedia(multimedia.getBbMultimediaId());
+        }
+        catch (SoapFaultException e) {
+            //See if the multimedia file actually exists
+            final List<BlackboardMultimediaResponse> userMultimedias = this.multimediaWSDao.getRepositoryMultimedias(null, multimedia.getBbMultimediaId(), null);
+            
+            //Multimedia file exists but we failed to remove it, throw the exception
+            if (!userMultimedias.isEmpty()) {
+                throw e;
+            }
+            
+            //Multimedia file doesn't exist on the BB side, remove our local DB version
+        }
+        this.multimediaDao.deleteMultimedia(multimedia);
     }
 
     @Override
@@ -174,10 +242,10 @@ public class SessionServiceImpl implements SessionService, ServletContextAware {
     public void createOrUpdateSession(ConferenceUser user, SessionForm sessionForm) {
         if (sessionForm.isNewSession()) {
             final BlackboardSessionResponse sessionResponse = sessionWSDao.createSession(user, sessionForm);
-            final String guestUrl = sessionWSDao.buildSessionUrl(sessionResponse.getSessionId(), "GUEST_PLACEHOLDER");
+            final String guestUrl = sessionWSDao.buildGuestSessionUrl(sessionResponse.getSessionId());
         	
         	//Remove guest username so that guest user's are prompted when they use the URL
-            sessionDao.createSession(sessionResponse, guestUrl.replace("&username=GUEST_PLACEHOLDER", ""));
+            sessionDao.createSession(sessionResponse, guestUrl);
         }
         else {
             final Session session = sessionDao.getSession(sessionForm.getSessionId());
@@ -190,7 +258,7 @@ public class SessionServiceImpl implements SessionService, ServletContextAware {
     @Transactional
     @PreAuthorize("hasRole('ROLE_ADMIN') || hasPermission(#sessionId, 'org.jasig.portlet.blackboardvcportlet.data.Session', 'edit')")
     public void addSessionChair(long sessionId, String displayName, String email) {
-        final ConferenceUser newSessionChair = this.conferenceUserService.getOrCreateConferenceUser(email, displayName);
+        final ConferenceUser newSessionChair = this.conferenceUserService.getOrCreateConferenceUser(displayName, email);
         
         final Session session = this.sessionDao.getSession(sessionId);
         final Set<ConferenceUser> sessionChairs = new LinkedHashSet<ConferenceUser>(this.getSessionChairs(session));
@@ -203,12 +271,12 @@ public class SessionServiceImpl implements SessionService, ServletContextAware {
     @Override
     @Transactional
     @PreAuthorize("hasRole('ROLE_ADMIN') || hasPermission(#sessionId, 'org.jasig.portlet.blackboardvcportlet.data.Session', 'edit')")
-    public void removeSessionChairs(long sessionId, String... emails) {
+    public void removeSessionChairs(long sessionId, long... userIds) {
         final Session session = this.sessionDao.getSession(sessionId);
         final Set<ConferenceUser> sessionChairs = new LinkedHashSet<ConferenceUser>(this.getSessionChairs(session));
         
-        for (final String email : emails) {
-            final ConferenceUser user = conferenceUserDao.getUser(email);
+        for (final long userId : userIds) {
+            final ConferenceUser user = conferenceUserDao.getUser(userId);
             if (user != null) {
                 sessionChairs.remove(user);
             }
@@ -222,7 +290,7 @@ public class SessionServiceImpl implements SessionService, ServletContextAware {
     @Transactional
     @PreAuthorize("hasRole('ROLE_ADMIN') || hasPermission(#sessionId, 'org.jasig.portlet.blackboardvcportlet.data.Session', 'edit')")
     public void addSessionNonChair(long sessionId, String displayName, String email) {
-        final ConferenceUser newSessionNonChair = this.conferenceUserService.getOrCreateConferenceUser(email, displayName);
+        final ConferenceUser newSessionNonChair = this.conferenceUserService.getOrCreateConferenceUser(displayName, email);
         
         final Session session = this.sessionDao.getSession(sessionId);
         final Set<ConferenceUser> sessionNonChairs = new LinkedHashSet<ConferenceUser>(this.getSessionNonChairs(session));
@@ -236,12 +304,12 @@ public class SessionServiceImpl implements SessionService, ServletContextAware {
     @Override
     @Transactional
     @PreAuthorize("hasRole('ROLE_ADMIN') || hasPermission(#sessionId, 'org.jasig.portlet.blackboardvcportlet.data.Session', 'edit')")
-    public void removeSessionNonChairs(long sessionId, String... emails) {
+    public void removeSessionNonChairs(long sessionId, long... userIds) {
         final Session session = this.sessionDao.getSession(sessionId);
         final Set<ConferenceUser> sessionNonChairs = new LinkedHashSet<ConferenceUser>(this.getSessionNonChairs(session));
         
-        for (final String email : emails) {
-            final ConferenceUser user = conferenceUserDao.getUser(email);
+        for (final long userId : userIds) {
+            final ConferenceUser user = conferenceUserDao.getUser(userId);
             if (user != null) {
                 sessionNonChairs.remove(user);
             }
@@ -265,7 +333,98 @@ public class SessionServiceImpl implements SessionService, ServletContextAware {
         final Multimedia multimedia = this.multimediaDao.createMultimedia(multimediaResponse, filename);
         
         //Associate Multimedia with session
-        this.sessionDao.addMultimediaToSession(session.getBbSessionId(), multimedia);
+        this.sessionDao.addMultimediaToSession(session, multimedia);
+    }
+    
+    @Override
+    @Transactional(noRollbackFor = { WebServiceClientException.class, XmlMappingException.class })
+    @PreAuthorize("hasRole('ROLE_ADMIN') || (hasRole('ROLE_FULL_ACCESS') && hasPermission(#sessionId, 'org.jasig.portlet.blackboardvcportlet.data.Session', 'edit'))")
+    public void deleteMultimedia(long sessionId, long... multimediaIds) {
+        final Session session = this.sessionDao.getSession(sessionId);
+        
+        final Set<Long> bbMultimediaIds = getBlackboardMultimediaIds(session);
+        for (final long multimediaId : multimediaIds) {
+            final Multimedia multimedia = this.multimediaDao.getMultimedia(multimediaId);
+            removeMultimediaFromSession(session, bbMultimediaIds, multimedia);
+        }
+    }
+    
+    @Override
+    @Transactional
+    @PreAuthorize("hasRole('ROLE_ADMIN') || (hasRole('ROLE_FULL_ACCESS') && hasPermission(#sessionId, 'org.jasig.portlet.blackboardvcportlet.data.Session', 'edit'))")
+    public void addPresentation(long sessionId, MultipartFile file) {
+        final Session session = this.sessionDao.getSession(sessionId);
+        if (session.getPresentation() != null) {
+            this.deletePresentation(session.getSessionId());
+        }
+        
+        final ConferenceUser conferenceUser = this.conferenceUserService.getCurrentConferenceUser();
+        
+        final BlackboardPresentationResponse presentationResponse = createSessionPresentation(session, conferenceUser, file);
+        
+        final String filename = FilenameUtils.getName(file.getOriginalFilename());
+        final Presentation presentation = this.presentationDao.createPresentation(presentationResponse, filename);
+        
+        //Associate Presentation with session
+        this.sessionDao.addPresentationToSession(session, presentation);
+    }
+    
+    @Override
+    @Transactional
+    @PreAuthorize("hasRole('ROLE_ADMIN') || (hasRole('ROLE_FULL_ACCESS') && hasPermission(#sessionId, 'org.jasig.portlet.blackboardvcportlet.data.Session', 'edit'))")
+    public void deletePresentation(long sessionId) {
+        final Session session = this.sessionDao.getSession(sessionId);
+        final Presentation presentation = session.getPresentation();
+        if (presentation == null) {
+            return;
+        }
+        
+        try {
+            this.presentationWSDao.deleteSessionPresenation(session.getBbSessionId(), presentation.getBbPresentationId());
+        }
+        catch (SoapFaultException e) {
+            //See if the presentation association actually exists
+            final List<BlackboardPresentationResponse> sessionPresentations = this.presentationWSDao.getSessionPresentations(session.getBbSessionId());
+            
+            //Session presentation exists but failed to remove it, throw the exception
+            if (!sessionPresentations.isEmpty()) {
+                throw e;
+            }
+            
+            //Presentation association doesn't exist but we still need to delete the association no our side.
+        }
+        this.sessionDao.removePresentationFromSession(session);
+        
+        
+        try {
+            this.presentationWSDao.deletePresentation(presentation.getBbPresentationId());
+        }
+        catch (SoapFaultException e) {
+            //See if the presentation actually exists
+            final List<BlackboardPresentationResponse> repositoryPresentations = this.presentationWSDao.getRepositoryPresentations(null, presentation.getBbPresentationId(), null);
+            
+            //Presentation exists but failed to remove it, throw the exception
+            if (!repositoryPresentations.isEmpty()) {
+                throw e;
+            }
+            
+            //Presentation doesn't exist but we still need to delete the association no our side.
+        }
+        this.presentationDao.deletePresentation(presentation);
+        
+        this.sessionDao.removePresentationFromSession(session);
+    }
+
+    private Set<Long> getBlackboardMultimediaIds(final Session session) {
+        final List<BlackboardMultimediaResponse> multimedias = this.multimediaWSDao.getSessionMultimedias(session.getBbSessionId());
+        final Iterator<BlackboardMultimediaResponse> multimediasItr = multimedias.iterator();
+        return ImmutableSet.copyOf(
+            Iterators.transform(multimediasItr, 
+                new Function<BlackboardMultimediaResponse, Long>() {
+                    public Long apply(BlackboardMultimediaResponse r) {
+                        return r.getMultimediaId();
+                    }
+                }));
     }
 
     private BlackboardMultimediaResponse createSessionMultimedia(Session session, ConferenceUser conferenceUser, MultipartFile file) {
@@ -280,10 +439,36 @@ public class SessionServiceImpl implements SessionService, ServletContextAware {
             //Upload the file to BB
             return this.multimediaWSDao.createSessionMultimedia(
                 session.getBbSessionId(), 
-                conferenceUser.getEmail(), 
+                conferenceUser.getUniqueId(), 
                 filename, 
                 "",
                 new DataHandler(new FileDataSource(multimediaFile)));
+        }
+        catch (IOException e) {
+            throw new RuntimeException("Failed to upload multimedia file '" + filename + "'", e);
+        }
+        finally {
+            FileUtils.deleteQuietly(multimediaFile);
+        }
+    }
+
+    //TODO consolidate these, extract uploadPresentation/createSessionMultimedia to generic interface and use that
+    private BlackboardPresentationResponse createSessionPresentation(Session session, ConferenceUser conferenceUser, MultipartFile file) {
+        final String filename = FilenameUtils.getName(file.getOriginalFilename());
+        
+        File multimediaFile = null;
+        try {
+            //Transfer the uploaded file to our own temp file so we can use a FileDataSource
+            multimediaFile = File.createTempFile(filename, ".tmp", this.tempDir);
+            file.transferTo(multimediaFile);
+            
+            //Upload the file to BB
+            return this.presentationWSDao.uploadPresentation(
+                    session.getBbSessionId(), 
+                    conferenceUser.getUniqueId(), 
+                    filename, 
+                    "",
+                    new DataHandler(new FileDataSource(multimediaFile)));
         }
         catch (IOException e) {
             throw new RuntimeException("Failed to upload multimedia file '" + filename + "'", e);
